Extract CLI IPC server out of DaemonServer into a module singleton
The AssistantIpcServer (CLI IPC) was a DaemonServer instance field. Move it into
an assistant-server module singleton with startCliIpcServer()/stopCliIpcServer(),
called from the daemon startup and shutdown sequences. startCliIpcServer()
preserves the EADDRINUSE re-throw that aborts startup when another daemon holds
the socket; other bind failures stay non-fatal. The isEaddrInUse helper moves to
assistant-server alongside it.
